vbs-server/requirements.txt
# VBS Simple Server - Minimal Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
motor>=3.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
# ML dependencies (optional - only needed for text search)
numpy>=1.24.0
torch>=2.0.0
transformers>=4.35.0
faiss-cpu>=1.7.4

sentencepiece>=0.2.1
timm==0.6.13
torchmetrics>=0.11.0
tensorboard>=2.13.0
tensorboardX>=2.6.0
torchscale>=0.3.0
fairscale==0.4.0
